refactor: rename test package from falsenegative to compositelit
