+++ goexhauerrors/analyzer_test.go
@@ -45,7 +45,7 @@ func TestAnalyzer(t *testing.T) {
 		"crosspkgdi/infra",
 		"crosspkgdi/usecase",
 		"ifacehigherorder",
-		"falsenegative",
+		"compositelit",
 	)
 }
